refactor(cpu-local): split register backends by architecture
