+++ components/cpu-local/src/lib.rs
@@ -11,6 +11,8 @@ extern crate std;
 
 mod header;
 mod identity;
+#[cfg(target_arch = "loongarch64")]
+pub mod loongarch64;
 pub mod platform;
 mod register;
 mod symbol;
@@ -35,24 +37,4 @@ pub mod raw {
         current_thread, get_task_pointer_raw as get_task_pointer, install_binding,
         set_task_pointer_raw as set_task_pointer,
     };
-}
-
-/// LoongArch host scratch-register assignments shared by trap and vCPU code.
-#[cfg(target_arch = "loongarch64")]
-pub mod loongarch64 {
-    /// Kernel stack pointer scratch slot used by exception entry.
-    pub const KSAVE_KSP: usize = 0;
-    /// First temporary-register scratch slot used by exception entry.
-    pub const KSAVE_T0: usize = 1;
-    /// Second temporary-register scratch slot used by exception entry.
-    pub const KSAVE_T1: usize = 2;
-    /// CPU-local runtime area-base shadow restored by exception entry.
-    pub const KSAVE_PERCPU: usize = 3;
-
-    /// Host CPU-local runtime area-base shadow.
-    pub const HOST_PERCPU_KS: usize = KSAVE_PERCPU;
-    /// Host stack scratch reserved for vCPU entry and exit.
-    pub const HOST_VCPU_KS: usize = 4;
-    /// Temporary scratch reserved for vCPU entry and exit.
-    pub const HOST_VCPU_TMP_KS: usize = 5;
 }
